Merge/master release

+++ src/SME.SGP.Aplicacao/Commands/Armazenamento/ExcluirArquivoFisico/ExcluirArquivoFisicoCommandHandler.cs
@@ -2,9 +2,7 @@
 using SME.SGP.Dominio;
 using SME.SGP.Infra;
 using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -21,9 +19,14 @@ namespace SME.SGP.Aplicacao
             var caminhoBase = ObterCaminhoArquivos(request.Tipo);
             var extencao = Path.GetExtension(request.Nome);
             var nomeArquivo = $"{request.Codigo}{extencao}";
-            var caminhoArquivo = Path.Combine($"{caminhoBase}", nomeArquivo).Replace(@"\", @"//");
+            var caminhoArquivo = $"{caminhoBase}//{nomeArquivo}".Replace(@"\", @"//");
 
-            File.Delete(caminhoArquivo);
+            if (File.Exists(caminhoArquivo))
+                File.Delete(caminhoArquivo);
+            else
+            {
+                var mensagem = $"Arquivo Informado para exclusão não existe no caminho {caminhoArquivo} ";
+            }
             return Task.FromResult(true);
         }
